Int 64 bit push where it does not affect speed

+++ src/densematrix.cpp
@@ -10,44 +10,44 @@ void densematrix::errorifempty(void)
     }
 }
 
-densematrix::densematrix(int numberofrows, int numberofcolumns)
+densematrix::densematrix(long long int numberofrows, long long int numberofcolumns)
 {
     numrows = numberofrows;
     numcols = numberofcolumns;
     myvalues = std::shared_ptr<double>(new double[numcols*numrows]);
 }
 
-densematrix::densematrix(int numberofrows, int numberofcolumns, double initvalue)
+densematrix::densematrix(long long int numberofrows, long long int numberofcolumns, double initvalue)
 {
     numrows = numberofrows;
     numcols = numberofcolumns;
     double* myvaluesptr = new double[numcols*numrows];
     
-    for (int i = 0; i < numcols*numrows; i++)
+    for (long long int i = 0; i < numcols*numrows; i++)
         myvaluesptr[i] = initvalue;
     
     myvalues = std::shared_ptr<double>(myvaluesptr);
 }
 
-densematrix::densematrix(int numberofrows, int numberofcolumns, const std::vector<double> valvec)
+densematrix::densematrix(long long int numberofrows, long long int numberofcolumns, const std::vector<double> valvec)
 {
     numrows = numberofrows;
     numcols = numberofcolumns;
     double* myvaluesptr = new double[numcols*numrows];
     
-    for (int i = 0; i < numcols*numrows; i++)
+    for (long long int i = 0; i < numcols*numrows; i++)
         myvaluesptr[i] = valvec[i];
     
     myvalues = std::shared_ptr<double>(myvaluesptr);
 }
 
-densematrix::densematrix(int numberofrows, int numberofcolumns, double init, double step)
+densematrix::densematrix(long long int numberofrows, long long int numberofcolumns, double init, double step)
 {
     numrows = numberofrows;
     numcols = numberofcolumns;
     double* myvaluesptr = new double[numcols*numrows];
     
-    for (int i = 0; i < numcols*numrows; i++)
+    for (long long int i = 0; i < numcols*numrows; i++)
         myvaluesptr[i] = init+i*step;
     
     myvalues = std::shared_ptr<double>(myvaluesptr);
@@ -61,7 +61,7 @@ densematrix::densematrix(std::vector<densematrix> input)
     // Calculate the final concatenated size:
     numcols = input[0].countcolumns();
     numrows = input[0].countrows();
-    for (int i = 1; i < input.size(); i++)
+    for (long long int i = 1; i < input.size(); i++)
     {
         numrows += input[i].countrows();
         if (input[i].countcolumns() != numcols)
@@ -72,11 +72,11 @@ densematrix::densematrix(std::vector<densematrix> input)
     }
     double* myvaluesptr = new double[numrows*numcols];
     
-    int index = 0;
-    for (int i = 0; i < input.size(); i++)
+    long long int index = 0;
+    for (long long int i = 0; i < input.size(); i++)
     {
         double* curvals = input[i].getvalues();
-        for (int j = 0; j < input[i].count(); j++)
+        for (long long int j = 0; j < input[i].count(); j++)
         {
             myvaluesptr[index] = curvals[j];
             index++;
@@ -86,10 +86,10 @@ densematrix::densematrix(std::vector<densematrix> input)
     myvalues = std::shared_ptr<double>(myvaluesptr);
 }
 
-void densematrix::setrow(int rownumber, std::vector<double> rowvals)
+void densematrix::setrow(long long int rownumber, std::vector<double> rowvals)
 {
     double* myvaluesptr = myvalues.get();
-    for (int i = 0; i < numcols; i++)
+    for (long long int i = 0; i < numcols; i++)
         myvaluesptr[rownumber*numcols+i] = rowvals[i];
 }
 
@@ -102,14 +102,14 @@ densematrix densematrix::flatten(void)
 }
 
 
-void densematrix::insert(int row, int col, densematrix toinsert)
+void densematrix::insert(long long int row, long long int col, densematrix toinsert)
 {
     double* myvaluesptr = myvalues.get();
     double* toinsertvaluesptr = toinsert.myvalues.get();
 
-    for (int i = 0; i < toinsert.numrows; i++)
+    for (long long int i = 0; i < toinsert.numrows; i++)
     {
-        for (int j = 0; j < toinsert.numcols; j++)
+        for (long long int j = 0; j < toinsert.numcols; j++)
             myvaluesptr[(row+i)*numcols+(col+j)] = toinsertvaluesptr[i*toinsert.numcols+j];
     }
 }
@@ -119,9 +119,9 @@ void densematrix::insertatrows(std::vector<int> selectedrows, densematrix toinse
     double* myvaluesptr = myvalues.get();
     double* toinsertvaluesptr = toinsert.myvalues.get();
     
-    for (int i = 0; i < selectedrows.size(); i++)
+    for (long long int i = 0; i < selectedrows.size(); i++)
     {
-        for (int col = 0; col < numcols; col++)
+        for (long long int col = 0; col < numcols; col++)
             myvaluesptr[selectedrows[i]*numcols+col] = toinsertvaluesptr[i*numcols+col];
     }
 }
@@ -131,20 +131,20 @@ void densematrix::insertatcolumns(std::vector<int> selectedcolumns, densematrix 
     double* myvaluesptr = myvalues.get();
     double* toinsertvaluesptr = toinsert.myvalues.get();
     
-    for (int row = 0; row < numrows; row++)
+    for (long long int row = 0; row < numrows; row++)
     {
-        for (int j = 0; j < selectedcolumns.size(); j++)
+        for (long long int j = 0; j < selectedcolumns.size(); j++)
             myvaluesptr[row*numcols+selectedcolumns[j]] = toinsertvaluesptr[row*toinsert.numcols+j];
     }
 }
 
-double densematrix::getvalue(int rownumber, int columnnumber)
+double densematrix::getvalue(long long int rownumber, long long int columnnumber)
 {
     double* myvaluesptr = myvalues.get();
     return myvaluesptr[rownumber*numcols+columnnumber];
 }
 
-void densematrix::setvalue(int rownumber, int columnnumber, double val)
+void densematrix::setvalue(long long int rownumber, long long int columnnumber, double val)
 {
     double* myvaluesptr = myvalues.get();
     myvaluesptr[rownumber*numcols+columnnumber] = val;
@@ -155,7 +155,7 @@ void densematrix::getvalues(std::vector<double>& topopulate)
     topopulate.resize(count());
 
     double* myvaluesptr = myvalues.get();
-    for (int i = 0; i < count(); i++)
+    for (long long int i = 0; i < count(); i++)
         topopulate[i] = myvaluesptr[i];
 }
 
@@ -169,7 +169,7 @@ densematrix densematrix::copy(void)
         densematrixcopy.myvalues = std::shared_ptr<double>(new double[numcols*numrows]);
         double* copiedmyvaluesptr = densematrixcopy.myvalues.get();
         double* myvaluesptr = myvalues.get();
-        for (int i = 0; i < numcols*numrows; i++)
+        for (long long int i = 0; i < numcols*numrows; i++)
             copiedmyvaluesptr[i] = myvaluesptr[i];
     }
         
@@ -181,9 +181,9 @@ void densematrix::print(void)
     printsize();
     
     double* myvaluesptr = myvalues.get();
-    for (int i = 0; i < numrows; i++)
+    for (long long int i = 0; i < numrows; i++)
     {
-        for (int j = 0; j < numcols; j++)
+        for (long long int j = 0; j < numcols; j++)
             std::cout << myvaluesptr[i*numcols+j] << " ";
         std::cout << std::endl;
     }
@@ -198,7 +198,7 @@ void densematrix::printsize(void)
 bool densematrix::isallzero(void)
 {
     double* myvaluesptr = myvalues.get();
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
     {
         if (myvaluesptr[i] != 0)
             return false;
@@ -214,7 +214,7 @@ void densematrix::transpose(void)
 densematrix densematrix::multiply(densematrix B)
 {
     // Number of rows and columns of A and B after transposition:
-    int numrowsA, numcolsA, numrowsB, numcolsB;
+    long long int numrowsA, numcolsA, numrowsB, numcolsB;
     if (not(istransposed))
     {
         numrowsA = numrows;
@@ -246,7 +246,7 @@ densematrix densematrix::multiply(densematrix B)
     
     
     // For too small matrices the overhead of BLAS is too visible and we use a homemade product.
-    int sizethreshold = 100*100;
+    long long int sizethreshold = 100*100;
     if (numrowsA*numcolsA > sizethreshold || numrowsB*numcolsB > sizethreshold)
     {
         // 'cblas_dgemm' computes alpha*A*B+beta*C and puts the result in C. Here we only compute A*B thus:
@@ -268,12 +268,12 @@ densematrix densematrix::multiply(densematrix B)
         double* Cmyvaluesptr = C.myvalues.get();
         
         // Hand coded matrix matrix product:
-        for (int m = 0; m < numrowsA; m++)
+        for (long long int m = 0; m < numrowsA; m++)
         {
-            for (int n = 0; n < numcolsB; n++)
+            for (long long int n = 0; n < numcolsB; n++)
             {
                 Cmyvaluesptr[m*numcolsB+n] = 0;
-                for (int k = 0; k < numcolsA; k++)
+                for (long long int k = 0; k < numcolsA; k++)
                 {
                     double valA, valB;
                     if (not(istransposed))
@@ -304,12 +304,12 @@ void densematrix::addproduct(double coef, densematrix B)
     
     if (coef == 1)
     {
-        for (int i = 0; i < numrows*numcols; i++)
+        for (long long int i = 0; i < numrows*numcols; i++)
             myvaluesptr[i] += Bmyvaluesptr[i];
     }
     else
     {
-        for (int i = 0; i < numrows*numcols; i++)
+        for (long long int i = 0; i < numrows*numcols; i++)
             myvaluesptr[i] += coef*Bmyvaluesptr[i];
     }
 }
@@ -322,12 +322,12 @@ densematrix densematrix::returnproduct(double coef)
     
     if (coef == 1)
     {
-        for (int i = 0; i < numrows*numcols; i++)
+        for (long long int i = 0; i < numrows*numcols; i++)
             outputmyvaluesptr[i] = myvaluesptr[i];
     }
     else
     {
-        for (int i = 0; i < numrows*numcols; i++)
+        for (long long int i = 0; i < numrows*numcols; i++)
             outputmyvaluesptr[i] = coef*myvaluesptr[i];
     }
     return output;
@@ -338,7 +338,7 @@ void densematrix::multiplyelementwise(densematrix B)
     double* myvaluesptr = myvalues.get();
     double* Bmyvaluesptr = B.myvalues.get();
     
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] *= Bmyvaluesptr[i];
 }
 
@@ -346,7 +346,7 @@ void densematrix::multiplyelementwise(double val)
 {
     double* myvaluesptr = myvalues.get();
     
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] *= val;
 }
 
@@ -355,7 +355,7 @@ void densematrix::add(densematrix B)
     double* myvaluesptr = myvalues.get();
     double* Bmyvaluesptr = B.myvalues.get();
     
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] += Bmyvaluesptr[i];
 }
 
@@ -364,7 +364,7 @@ void densematrix::subtract(densematrix B)
     double* myvaluesptr = myvalues.get();
     double* Bmyvaluesptr = B.myvalues.get();
     
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] -= Bmyvaluesptr[i];
 }
 
@@ -372,7 +372,7 @@ void densematrix::minus(void)
 {
     double* myvaluesptr = myvalues.get();
     
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = -myvaluesptr[i];
 }
 
@@ -381,7 +381,7 @@ void densematrix::power(densematrix exponent)
     double* myvaluesptr = myvalues.get();
     double* expmyvaluesptr = exponent.myvalues.get();
     
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
             myvaluesptr[i] = std::pow(myvaluesptr[i], expmyvaluesptr[i]);
 }
 
@@ -389,7 +389,7 @@ void densematrix::invert(void)
 {
     double* myvaluesptr = myvalues.get();
     
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = 1.0/myvaluesptr[i];
 }
 
@@ -397,7 +397,7 @@ void densematrix::abs(void)
 {
     double* myvaluesptr = myvalues.get();
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = std::abs(myvaluesptr[i]);
 }
 
@@ -405,7 +405,7 @@ void densematrix::sin(void)
 {
     double* myvaluesptr = myvalues.get();
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = std::sin(myvaluesptr[i]);
 }
 
@@ -413,7 +413,7 @@ void densematrix::cos(void)
 {
     double* myvaluesptr = myvalues.get();
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = std::cos(myvaluesptr[i]);
 }
 
@@ -421,7 +421,7 @@ void densematrix::tan(void)
 {
     double* myvaluesptr = myvalues.get();
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = std::tan(myvaluesptr[i]);
 }
 
@@ -429,7 +429,7 @@ void densematrix::asin(void)
 {
     double* myvaluesptr = myvalues.get();
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = std::asin(myvaluesptr[i]);
 }
 
@@ -437,7 +437,7 @@ void densematrix::acos(void)
 {
     double* myvaluesptr = myvalues.get();
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = std::acos(myvaluesptr[i]);
 }
 
@@ -445,7 +445,7 @@ void densematrix::atan(void)
 {
     double* myvaluesptr = myvalues.get();
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = std::atan(myvaluesptr[i]);
 }
 
@@ -453,7 +453,7 @@ void densematrix::log10(void)
 {
     double* myvaluesptr = myvalues.get();
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
         myvaluesptr[i] = std::log10(myvaluesptr[i]);
 }
 
@@ -465,7 +465,7 @@ double densematrix::max(void)
     
     double val = myvaluesptr[0];
 
-    for (int i = 1; i < numrows*numcols; i++)
+    for (long long int i = 1; i < numrows*numcols; i++)
     {
         if (myvaluesptr[i] > val)
             val = myvaluesptr[i];
@@ -481,7 +481,7 @@ double densematrix::maxabs(void)
     
     double val = std::abs(myvaluesptr[0]);
 
-    for (int i = 1; i < numrows*numcols; i++)
+    for (long long int i = 1; i < numrows*numcols; i++)
     {
         if (std::abs(myvaluesptr[i]) > val)
             val = std::abs(myvaluesptr[i]);
@@ -496,7 +496,7 @@ double densematrix::sum(void)
     double* myvaluesptr = myvalues.get();
     double val = 0;
 
-    for (int i = 0; i < numrows*numcols; i++)
+    for (long long int i = 0; i < numrows*numcols; i++)
             val += myvaluesptr[i];
     return val;
 }
@@ -505,9 +505,9 @@ void densematrix::multiplycolumns(std::vector<double> input)
 {
     double* myvaluesptr = myvalues.get();
     
-    for (int i = 0; i < numrows; i++)
+    for (long long int i = 0; i < numrows; i++)
     {
-        for (int j = 0; j < numcols; j++)
+        for (long long int j = 0; j < numcols; j++)
             myvaluesptr[i*numcols + j] *= input[j];
     }
 }
@@ -520,11 +520,11 @@ densematrix densematrix::multiplyallrows(densematrix input)
     double* inmyvaluesptr = input.myvalues.get();
     double* outmyvaluesptr = output.myvalues.get();
     
-    for (int i = 0; i < numrows; i++)
+    for (long long int i = 0; i < numrows; i++)
     {
-        for (int j = 0; j < input.numrows; j++)
+        for (long long int j = 0; j < input.numrows; j++)
         {
-            for (int k = 0; k < numcols; k++)
+            for (long long int k = 0; k < numcols; k++)
                 outmyvaluesptr[(i*input.numrows + j)*numcols + k] = myvaluesptr[i*numcols+k] * inmyvaluesptr[j*input.numcols+k];
         }
     }
@@ -540,8 +540,8 @@ densematrix densematrix::duplicatevertically(int n)
     
     for (int duplicate = 0; duplicate < n; duplicate++)
     {
-        int offset = duplicate*numrows*numcols;
-        for (int i = 0; i < numrows*numcols; i++)
+        long long int offset = duplicate*numrows*numcols;
+        for (long long int i = 0; i < numrows*numcols; i++)
             outmyvaluesptr[offset + i] = myvaluesptr[i];
     }
     
@@ -550,16 +550,16 @@ densematrix densematrix::duplicatevertically(int n)
 
 densematrix densematrix::extractrows(std::vector<int> selected)
 {
-    int numselected = selected.size();
+    long long int numselected = selected.size();
 
     densematrix output(numselected, numcols);
     
     double* vals = getvalues();
     double* outvals = output.getvalues();
-    for (int i = 0; i < numselected; i++)
+    for (long long int i = 0; i < numselected; i++)
     {
-        int row = selected[i];
-        for (int j = 0; j < numcols; j++)
+        long long int row = selected[i];
+        for (long long int j = 0; j < numcols; j++)
             outvals[i*numcols+j] = vals[row*numcols+j];
     }
     
@@ -568,17 +568,17 @@ densematrix densematrix::extractrows(std::vector<int> selected)
 
 densematrix densematrix::extractcols(std::vector<int> selected)
 {
-    int numselected = selected.size();
+    long long int numselected = selected.size();
     
     densematrix output(numrows, numselected);
     
     double* vals = getvalues();
     double* outvals = output.getvalues();
-    for (int i = 0; i < numrows; i++)
+    for (long long int i = 0; i < numrows; i++)
     {
-        for (int j = 0; j < numselected; j++)
+        for (long long int j = 0; j < numselected; j++)
         {
-            int col = selected[j];
+            long long int col = selected[j];
             outvals[i*numselected+j] = vals[i*numcols+col];
         }
     }
@@ -586,19 +586,19 @@ densematrix densematrix::extractcols(std::vector<int> selected)
     return output;
 }
 
-densematrix densematrix::extractrows(int rangebegin, int rangeend)
+densematrix densematrix::extractrows(long long int rangebegin, long long int rangeend)
 {
     std::vector<int> selected(rangeend-rangebegin+1);
-    for (int i = 0; i < selected.size(); i++)
+    for (long long int i = 0; i < selected.size(); i++)
         selected[i] = rangebegin + i;
         
     return extractrows(selected);
 }
 
-densematrix densematrix::extractcols(int rangebegin, int rangeend)
+densematrix densematrix::extractcols(long long int rangebegin, long long int rangeend)
 {
     std::vector<int> selected(rangeend-rangebegin+1);
-    for (int i = 0; i < selected.size(); i++)
+    for (long long int i = 0; i < selected.size(); i++)
         selected[i] = rangebegin + i;
         
     return extractcols(selected);
